app/requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
